Configuration Merging
Spent half of my coach back from Lithuania ironing this out, proper pain in the ass but got there in the end.

Essentially, I wanted to allow the user to modify a config, defining their own options when generating files, however, if they modified the config and removed properties or added additional ones, I didn't want these to be passed into the templates.

So what i've done, is by default we create a new SettingsContainer (which holds a default config), then attempt to import the user defined one. We loop through every key, if nested, we use a recursive function. If the key exists in the template, then replace, else ignore the key as it doesn't exist in our template.

Controller template now shows example of user defines async option.